Move ME biomes
